+++ src/api/user/user.repository.ts
@@ -1,12 +1,12 @@
 import { Repository } from 'typeorm';
 
-import { CustomRepository } from '@/common/customRepository.decorator';
-import { CreateUserRetDto } from '@/dto/user/create-user.dto';
+import { CustomRepository } from '@/common/decorators';
+import { UserCreateDto } from '@/dto/user';
 import { User } from '@/entity/user.entity';
 
 @CustomRepository(User)
 export class UserRepository extends Repository<User> {
-  async createUser(userInfo: CreateUserRetDto) {
+  async createUser(userInfo: UserCreateDto) {
     return this.save(this.create(userInfo));
   }
